Улучшения UI, поиск по тикетам и правки CI
- Поиск по тикетам: debounced FTS через ?search= (400ms, серверная сторона)
- Сайдбар: сворачивается до 48px с кнопкой [›] для разворота
- Вход перенаправляет на /dashboard вместо /chat
- Колокольчик уведомлений перемещён левее имени пользователя, иконка stroke=1.5
- Тренды и «Ваш месяц»: навигатор по месяцам [‹] Май 2026 [›]
- Backend /stats/trends: параметры since/until вместо period_days
- Экспорт CSV доступен только admin и agent
- CI: ruff + mypy — исправлены I001, F401, F841, no-any-return, форматирование

+++ backend/app/services/quality_signals.py
@@ -97,7 +97,7 @@ def _decay_weight(age_days: float, half_life: float = FEEDBACK_DECAY_HALF_LIFE_D
     Используем 2** вместо math.exp() ради читаемости: half-life явно ровно
     FEEDBACK_DECAY_HALF_LIFE_DAYS, а не λ из e^(-λt).
     """
-    return 2.0 ** (-max(0.0, age_days) / half_life)
+    return float(2.0 ** (-max(0.0, age_days) / half_life))
 
 
 # ── Вычисление grade ────────────────────────────────────────────────────────
